--HG--
branch : next


Former-commit-id: 4e4933babf91ada47a1d7161daefb8426b06927f

+++ RavenQuestionnaire.Web/Core.Supervisor/Views/Interview/InterviewLinkedQuestionView.cs
@@ -12,7 +12,7 @@ namespace Core.Supervisor.Views.Interview
     {
         public InterviewLinkedQuestionView(IQuestion question, InterviewQuestion answeredQuestion,
             Dictionary<Guid, string> variablesMap, Dictionary<string, string> answersForTitleSubstitution,
-            Func<Guid, Dictionary<int[], string>> getAvailableOptions, bool isParentGroupDisabled)
+            Func<Guid, Dictionary<decimal[], string>> getAvailableOptions, bool isParentGroupDisabled)
             : base(question, answeredQuestion, variablesMap, answersForTitleSubstitution, isParentGroupDisabled)
         {
             this.Options = getAvailableOptions(question.PublicKey).Select(a => new QuestionOptionView
