+++ src/Core/BoundedContexts/Supervisor/WB.Core.BoundedContexts.Supervisor/Views/DataExport/InterviewDataExportRerord.cs
@@ -8,7 +8,7 @@ namespace WB.Core.BoundedContexts.Supervisor.Views.DataExport
 {
     public class InterviewDataExportRerord
     {
-        public InterviewDataExportRerord(Guid interviewId, int recordId, int? parentLevelId, Dictionary<Guid, ExportedQuestion> questions)
+        public InterviewDataExportRerord(Guid interviewId, int recordId, decimal? parentLevelId, Dictionary<Guid, ExportedQuestion> questions)
         {
             InterviewId = interviewId;
             RecordId = recordId;
@@ -18,7 +18,7 @@ namespace WB.Core.BoundedContexts.Supervisor.Views.DataExport
 
         public Guid InterviewId { get; private set; }
         public int RecordId { get; private set; }
-        public int? ParentRecordId { get; private set; }
+        public decimal? ParentRecordId { get; private set; }
         public Dictionary<Guid,ExportedQuestion> Questions { get; private set; }
     }
 }
