+++ src/Core/SharedKernels/DataCollection/WB.Core.SharedKernels.DataCollection/Commands/Interview/AnswerDateTimeQuestionCommand.cs
@@ -9,7 +9,7 @@ namespace WB.Core.SharedKernels.DataCollection.Commands.Interview
     {
         public DateTime Answer { get; private set; }
 
-        public AnswerDateTimeQuestionCommand(Guid interviewId, Guid userId, Guid questionId, int[] rosterVector, DateTime answerTime, DateTime answer)
+        public AnswerDateTimeQuestionCommand(Guid interviewId, Guid userId, Guid questionId, decimal[] rosterVector, DateTime answerTime, DateTime answer)
             : base(interviewId, userId, questionId, rosterVector, answerTime)
         {
             this.Answer = answer;
