+++ src/Core/SharedKernels/DataCollection/WB.Core.SharedKernels.DataCollection/Commands/Interview/AnswerSingleOptionLinkedQuestionCommand.cs
@@ -7,9 +7,9 @@ namespace WB.Core.SharedKernels.DataCollection.Commands.Interview
     [MapsToAggregateRootMethod(typeof(Implementation.Aggregates.Interview), "AnswerSingleOptionLinkedQuestion")]
     public class AnswerSingleOptionLinkedQuestionCommand : AnswerQuestionCommand
     {
-        public int[] SelectedPropagationVector { get; private set; }
+        public decimal[] SelectedPropagationVector { get; private set; }
 
-        public AnswerSingleOptionLinkedQuestionCommand(Guid interviewId, Guid userId, Guid questionId, int[] rosterVector, DateTime answerTime, int[] selectedPropagationVector)
+        public AnswerSingleOptionLinkedQuestionCommand(Guid interviewId, Guid userId, Guid questionId, decimal[] rosterVector, DateTime answerTime, decimal[] selectedPropagationVector)
             : base(interviewId, userId, questionId, rosterVector, answerTime)
         {
             this.SelectedPropagationVector = selectedPropagationVector;
