+++ src/Core/SharedKernels/DataCollection/WB.Core.SharedKernels.DataCollection/Events/Interview/Base/QuestionActiveEvent.cs
@@ -5,9 +5,9 @@ namespace WB.Core.SharedKernels.DataCollection.Events.Interview.Base
     public abstract class QuestionActiveEvent : InterviewActiveEvent
     {
         public Guid QuestionId { get; private set; }
-        public int[] PropagationVector { get; private set; }
+        public decimal[] PropagationVector { get; private set; }
 
-        protected QuestionActiveEvent(Guid userId, Guid questionId, int[] propagationVector)
+        protected QuestionActiveEvent(Guid userId, Guid questionId, decimal[] propagationVector)
             : base(userId)
         {
             this.QuestionId = questionId;
