+++ src/Core/SharedKernels/DataCollection/WB.Core.SharedKernels.DataCollection/Events/Interview/NumericRealQuestionAnswered.cs
@@ -7,7 +7,7 @@ namespace WB.Core.SharedKernels.DataCollection.Events.Interview
     {
         public decimal Answer { get; private set; }
 
-        public NumericRealQuestionAnswered(Guid userId, Guid questionId, int[] propagationVector, DateTime answerTime, decimal answer)
+        public NumericRealQuestionAnswered(Guid userId, Guid questionId, decimal[] propagationVector, DateTime answerTime, decimal answer)
             : base(userId, questionId, propagationVector, answerTime)
         {
             this.Answer = answer;
